fix: correct limit calculation in movies list response
- Updated the limit calculation in the movies list response to use the total directly instead of casting it to i64, ensuring accurate pagination behavior.

+++ crates/server/src/routes/movies.rs
@@ -68,7 +68,7 @@ pub async fn list(
         Json(json!({
             "items": movies,
             "total": total,
-            "limit": limit.unwrap_or(total as i64),
+            "limit": limit.unwrap_or(total),
             "offset": offset
         })),
     ))
